更新atvoss模板功能 & 修改atvoss源码地址

Created-by: wulinyu
Commit-by: wulinyu
Merged-by: turing_project1
Description: ## 描述
<!--在这里详细描述你的改动，包括改动的原因和所采取的方法。-->
1. atvoss 更新模板
2. 修改atvoss源码地址

## 关联的Issue
<!-- 如果这个PR是为了解决特定的Issue，请在这里提供Issue链接。例如：关联Issue #123-->
<!-- 如果这个PR是为了解决特定的问题单，请在这里描述问题单单号。-->

## 测试
<!--描述进行了哪些测试来验证你的改动。包括但不限于二级冒烟、算子泛化等。-->

## 文档更新
<!--如果这个PR包含文档的更新，请在这里指出。例如：更新了README.md文件。-->

## 类型标签
<!--  [x] 表示选中 -->
- [ ] Bug修复
- [x] 新特性
- [ ] 性能优化
- [ ] 文档更新
- [ ] 其他，请描述：

See merge request: cann/atvoss!3

+++ scripts/package/atvos/scripts/atvos_uninstall.sh
@@ -176,7 +176,7 @@ remove_module() {
   local remain_dir_list=$(find ${TARGET_MOULDE_DIR} -mindepth 1 -maxdepth 1 -type d)
   for remain_dir in ${remain_dir_list}; do
     if [ "$(find "${remain_dir}" -type f 2>&1)" = "" ]; then
-      rm -rf ${remain_dir}
+      [ -n "${remain_dir}" ] && rm -rf ${remain_dir}
     fi
   done
 }
